+++ mini_udp/src/context.rs
@@ -14,11 +14,13 @@ use crate::prelude::*;
 /// type Message = String;
 ///
 /// impl MiniUdpContext for UdpCtx {
-///     type SEND = Message;
-///     type RECV = Message;
+///     type Send = Message;
+///     type Recv = Message;
 ///     const PROTOCOL_VERSION: u32 = 0;
 ///
-///     type REVERSE = UdpContext<Self::RECV, Self::SEND, { Self::PROTOCOL_VERSION }>;
+///     type Reverse = UdpContext<Self::Recv, Self::Send, { Self::PROTOCOL_VERSION }>;
+///     
+///     type ErrorHandling = mini_udp::context::error_handlers::WarnOnError;
 /// }
 ///
 /// let _com = UdpCommunicator::<UdpCtx>::default();
@@ -35,27 +37,66 @@ use crate::prelude::*;
 /// let _com = UdpCommunicator::<UdpCtx>::default();
 /// ```
 pub trait MiniUdpContext {
-    type SEND: ByteRepr;
-    type RECV: ByteRepr;
+    type Send: ByteRepr;
+    type Recv: ByteRepr;
     const PROTOCOL_VERSION: u32;
 
-    /// Reverse the `SEND` and `RECV` parameters, to represent the context of the connected
+    /// Reverse the `Send` and `Recv` parameters, to represent the context of the connected
     /// communicator instead (which should receive what this sends, and send what this receives).
-    type REVERSE: MiniUdpContext;
+    type Reverse: MiniUdpContext<Send = Self::Recv, Recv = Self::Send, ErrorHandling = Self::ErrorHandling>;
+
+    type ErrorHandling: ErrorHandlingStrategy;
 }
 
 /// See the trait docs for [`MiniUdpContext`].
-pub struct UdpContext<SEND, RECV, const PROTOCOL_VERSION: u32> {
+pub struct UdpContext<
+    SEND,
+    RECV,
+    const PROTOCOL_VERSION: u32,
+    ErrorHandling = error_handlers::WarnOnError,
+> {
     _send: PhantomData<SEND>,
     _recv: PhantomData<RECV>,
+    _error_handling: PhantomData<ErrorHandling>,
 }
 
-impl<SEND: ByteRepr, RECV: ByteRepr, const PROTOCOL_VERSION: u32> MiniUdpContext
-    for UdpContext<SEND, RECV, PROTOCOL_VERSION>
+impl<
+    Send: ByteRepr,
+    Recv: ByteRepr,
+    const PROTOCOL_VERSION: u32,
+    ErrorHandling: ErrorHandlingStrategy,
+> MiniUdpContext for UdpContext<Send, Recv, PROTOCOL_VERSION, ErrorHandling>
 {
-    type SEND = SEND;
-    type RECV = RECV;
+    type Send = Send;
+    type Recv = Recv;
     const PROTOCOL_VERSION: u32 = PROTOCOL_VERSION;
 
-    type REVERSE = UdpContext<RECV, SEND, PROTOCOL_VERSION>;
+    type Reverse = UdpContext<Recv, Send, PROTOCOL_VERSION, ErrorHandling>;
+
+    type ErrorHandling = ErrorHandling;
+}
+
+pub trait ErrorHandlingStrategy {
+    type Handler;
+    fn handle_error(handler: &mut Self::Handler, error: Error);
+}
+
+pub mod error_handlers {
+    use crate::prelude::*;
+
+    pub struct WarnOnError;
+    impl ErrorHandlingStrategy for WarnOnError {
+        type Handler = ();
+        fn handle_error(_handler: &mut Self::Handler, error: Error) {
+            warn!("{error}");
+        }
+    }
+
+    pub struct PanicOnError;
+    impl ErrorHandlingStrategy for PanicOnError {
+        type Handler = ();
+        fn handle_error(_handler: &mut Self::Handler, error: Error) {
+            panic!("{error}");
+        }
+    }
 }