+++ mini_udp/src/lib.rs
@@ -49,7 +49,7 @@
 //! /// that was send from a communicator with a different version, the CRC check will fail.
 //! const PROTOCOL_VERSION: u32 = 1;
 //! type ClientCtx = UdpContext<MessageToServer, MessageToClient, PROTOCOL_VERSION>;
-//! type ServerCtx = <ClientCtx as MiniUdpContext>::REVERSE;
+//! type ServerCtx = <ClientCtx as MiniUdpContext>::Reverse;
 //!
 //! const POSITION: [f32; 3] = [-1., 0.004, 2482.3];
 //!
@@ -107,7 +107,10 @@ mod byte_repr;
 pub use byte_repr::{ByteRepr, ByteReprError, StaticByteRepr};
 
 pub mod communicator;
-mod context;
+pub mod context;
+mod error;
+pub use error::Error;
+
 pub mod packet;
 mod packet_ack;
 pub mod prelude;